Get app ui information from script

Get app ui information from script

get_file_ui_definition_info(file_lines, type = "single-file")

Arguments

file_lines

Character vector of the lines of the file that defines a shiny app's ui (as from readLines()).

type

Is the app a single-file app? E.g. app is container entirely in app.R? Or is it multi-file?

Value

List with both the type and file_lines mirrored from the arguments of the same name. Along with this the ui_bounds containing the start and end lines of the file section that defines the app's ui, the ui_tree IR that defines that ui, and loaded_libraries: libraries loaded via library() calls in the script.

Examples

# Note the use of the triple colon, this function is not exported
# shinyuieditor:::get_file_ui_definition_info(...)

# Can handle single-file app.R
app_loc <- system.file(
  "app-templates/geyser_single-file/app.R",
  package = "shinyuieditor"
 )
shinyuieditor:::get_file_ui_definition_info(readLines(app_loc), type = "single-file")
#> $type
#> [1] "single-file"
#> 
#> $file_lines
#>  [1] "library(shiny)"                                                 
#>  [2] "library(gridlayout)"                                            
#>  [3] "library(gt)"                                                    
#>  [4] ""                                                               
#>  [5] "# App template from the shinyuieditor"                          
#>  [6] "ui <- grid_page("                                               
#>  [7] "  layout = c("                                                  
#>  [8] "    \"header header\","                                         
#>  [9] "    \"sidebar distPlot\","                                      
#> [10] "    \"sidebar bluePlot\""                                       
#> [11] "  ),"                                                           
#> [12] "  row_sizes = c("                                               
#> [13] "    \"100px\","                                                 
#> [14] "    \"1fr\","                                                   
#> [15] "    \"1fr\""                                                    
#> [16] "  ),"                                                           
#> [17] "  col_sizes = c("                                               
#> [18] "    \"250px\","                                                 
#> [19] "    \"1fr\""                                                    
#> [20] "  ),"                                                           
#> [21] "  gap_size = \"1rem\","                                         
#> [22] "  grid_card("                                                   
#> [23] "    area = \"sidebar\","                                        
#> [24] "    item_gap = \"12px\","                                       
#> [25] "    item_alignment = \"top\","                                  
#> [26] "    title = \"Settings\","                                      
#> [27] "    sliderInput("                                               
#> [28] "      inputId = \"bins\","                                      
#> [29] "      label = \"Number of Bins\","                              
#> [30] "      min = 12L,"                                               
#> [31] "      max = 100L,"                                              
#> [32] "      value = 30L,"                                             
#> [33] "      width = \"100%\""                                         
#> [34] "    )"                                                          
#> [35] "  ),"                                                           
#> [36] "  grid_card_text("                                              
#> [37] "    area = \"header\","                                         
#> [38] "    content = \"Geysers!\","                                    
#> [39] "    alignment = \"start\","                                     
#> [40] "    is_title = FALSE"                                           
#> [41] "  ),"                                                           
#> [42] "  grid_card_plot(area = \"bluePlot\"),"                         
#> [43] "  grid_card_plot(area = \"distPlot\")"                          
#> [44] ")"                                                              
#> [45] ""                                                               
#> [46] "# Define server logic required to draw a histogram"             
#> [47] "server <- function(input, output) {"                            
#> [48] ""                                                               
#> [49] "  output$distPlot <- renderPlot({"                              
#> [50] "    # generate bins based on input$bins from ui.R"              
#> [51] "    x    <- faithful[, 2]"                                      
#> [52] "    bins <- seq(min(x), max(x), length.out = input$bins + 1)"   
#> [53] ""                                                               
#> [54] "    # draw the histogram with the specified number of bins"     
#> [55] "    hist(x, breaks = bins, col = 'darkgray', border = 'white')" 
#> [56] "  })"                                                           
#> [57] ""                                                               
#> [58] "  output$bluePlot <- renderPlot({"                              
#> [59] "    # generate bins based on input$bins from ui.R"              
#> [60] "    x    <- faithful[, 2]"                                      
#> [61] "    bins <- seq(min(x), max(x), length.out = input$bins + 1)"   
#> [62] ""                                                               
#> [63] "    # draw the histogram with the specified number of bins"     
#> [64] "    hist(x, breaks = bins, col = 'steelblue', border = 'white')"
#> [65] "  })"                                                           
#> [66] ""                                                               
#> [67] "}"                                                              
#> [68] ""                                                               
#> [69] "shinyApp(ui, server)"                                           
#> [70] ""                                                               
#> 
#> $ui_bounds
#> $ui_bounds$start
#> [1] 6
#> 
#> $ui_bounds$end
#> [1] 44
#> 
#> 
#> $ui_tree
#> $ui_tree$uiName
#> [1] "gridlayout::grid_page"
#> 
#> $ui_tree$uiArguments
#> $ui_tree$uiArguments$row_sizes
#> [1] "100px" "1fr"   "1fr"  
#> 
#> $ui_tree$uiArguments$col_sizes
#> [1] "250px" "1fr"  
#> 
#> $ui_tree$uiArguments$gap_size
#> [1] "1rem"
#> 
#> $ui_tree$uiArguments$areas
#>      [,1]      [,2]      
#> [1,] "header"  "header"  
#> [2,] "sidebar" "distPlot"
#> [3,] "sidebar" "bluePlot"
#> 
#> 
#> $ui_tree$uiChildren
#> $ui_tree$uiChildren[[1]]
#> $ui_tree$uiChildren[[1]]$uiName
#> [1] "gridlayout::grid_card"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments
#> $ui_tree$uiChildren[[1]]$uiArguments$area
#> [1] "sidebar"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments$item_alignment
#> [1] "top"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments$title
#> [1] "Settings"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments$item_gap
#> [1] "12px"
#> 
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiName
#> [1] "shiny::sliderInput"
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments$inputId
#> [1] "bins"
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments$label
#> [1] "Number of Bins"
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments$min
#> [1] 12
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments$max
#> [1] 100
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments$value
#> [1] 30
#> 
#> $ui_tree$uiChildren[[1]]$uiChildren[[1]]$uiArguments$width
#> [1] "100%"
#> 
#> 
#> 
#> 
#> 
#> $ui_tree$uiChildren[[2]]
#> $ui_tree$uiChildren[[2]]$uiName
#> [1] "gridlayout::grid_card_text"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments
#> $ui_tree$uiChildren[[2]]$uiArguments$area
#> [1] "header"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments$content
#> [1] "Geysers!"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments$alignment
#> [1] "start"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments$is_title
#> [1] FALSE
#> 
#> 
#> 
#> $ui_tree$uiChildren[[3]]
#> $ui_tree$uiChildren[[3]]$uiName
#> [1] "gridlayout::grid_card_plot"
#> 
#> $ui_tree$uiChildren[[3]]$uiArguments
#> $ui_tree$uiChildren[[3]]$uiArguments$area
#> [1] "bluePlot"
#> 
#> 
#> 
#> $ui_tree$uiChildren[[4]]
#> $ui_tree$uiChildren[[4]]$uiName
#> [1] "gridlayout::grid_card_plot"
#> 
#> $ui_tree$uiChildren[[4]]$uiArguments
#> $ui_tree$uiChildren[[4]]$uiArguments$area
#> [1] "distPlot"
#> 
#> 
#> 
#> 
#> 
#> $loaded_libraries
#> [1] "shiny"      "gridlayout" "gt"        
#> 

# Also handles multi-file apps
app_loc <- system.file("app-templates/geyser/ui.R", package = "shinyuieditor")
shinyuieditor:::get_file_ui_definition_info(readLines(app_loc), type = "multi-file")
#> $type
#> [1] "multi-file"
#> 
#> $file_lines
#>  [1] "gridlayout::grid_page("                            
#>  [2] "  layout = c("                                     
#>  [3] "    \"header header\","                            
#>  [4] "    \"sidebar bluePlot\","                         
#>  [5] "    \"sidebar distPlot\""                          
#>  [6] "  ),"                                              
#>  [7] "  row_sizes = c("                                  
#>  [8] "    \"100px\","                                    
#>  [9] "    \"1fr\","                                      
#> [10] "    \"1fr\""                                       
#> [11] "  ),"                                              
#> [12] "  col_sizes = c("                                  
#> [13] "    \"250px\","                                    
#> [14] "    \"1fr\""                                       
#> [15] "  ),"                                              
#> [16] "  gap_size = \"15px\","                            
#> [17] "  gridlayout::grid_card_text("                     
#> [18] "    area = \"header\","                            
#> [19] "    content = \"Geysers!\","                       
#> [20] "    h_align = \"start\","                          
#> [21] "    is_title = TRUE"                               
#> [22] "  ),"                                              
#> [23] "  gridlayout::grid_card("                          
#> [24] "    area = \"sidebar\","                           
#> [25] "    item_alignment = \"top\","                     
#> [26] "    item_gap = \"12px\","                          
#> [27] "    shiny::sliderInput("                           
#> [28] "      inputId = \"bins\","                         
#> [29] "      label = \"Number of Bins\","                 
#> [30] "      min = 5L,"                                   
#> [31] "      max = 50L,"                                  
#> [32] "      value = 20L,"                                
#> [33] "      width = \"100%\""                            
#> [34] "    )"                                             
#> [35] "  ),"                                              
#> [36] "  gridlayout::grid_card_plot(area = \"distPlot\"),"
#> [37] "  gridlayout::grid_card_plot(area = \"bluePlot\")" 
#> [38] ")"                                                 
#> 
#> $ui_bounds
#> $ui_bounds$start
#> [1] 1
#> 
#> $ui_bounds$end
#> [1] 38
#> 
#> 
#> $ui_tree
#> $ui_tree$uiName
#> [1] "gridlayout::grid_page"
#> 
#> $ui_tree$uiArguments
#> $ui_tree$uiArguments$row_sizes
#> [1] "100px" "1fr"   "1fr"  
#> 
#> $ui_tree$uiArguments$col_sizes
#> [1] "250px" "1fr"  
#> 
#> $ui_tree$uiArguments$gap_size
#> [1] "15px"
#> 
#> $ui_tree$uiArguments$areas
#>      [,1]      [,2]      
#> [1,] "header"  "header"  
#> [2,] "sidebar" "bluePlot"
#> [3,] "sidebar" "distPlot"
#> 
#> 
#> $ui_tree$uiChildren
#> $ui_tree$uiChildren[[1]]
#> $ui_tree$uiChildren[[1]]$uiName
#> [1] "gridlayout::grid_card_text"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments
#> $ui_tree$uiChildren[[1]]$uiArguments$area
#> [1] "header"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments$content
#> [1] "Geysers!"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments$h_align
#> [1] "start"
#> 
#> $ui_tree$uiChildren[[1]]$uiArguments$is_title
#> [1] TRUE
#> 
#> 
#> 
#> $ui_tree$uiChildren[[2]]
#> $ui_tree$uiChildren[[2]]$uiName
#> [1] "gridlayout::grid_card"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments
#> $ui_tree$uiChildren[[2]]$uiArguments$area
#> [1] "sidebar"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments$item_alignment
#> [1] "top"
#> 
#> $ui_tree$uiChildren[[2]]$uiArguments$item_gap
#> [1] "12px"
#> 
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiName
#> [1] "shiny::sliderInput"
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments$inputId
#> [1] "bins"
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments$label
#> [1] "Number of Bins"
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments$min
#> [1] 5
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments$max
#> [1] 50
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments$value
#> [1] 20
#> 
#> $ui_tree$uiChildren[[2]]$uiChildren[[1]]$uiArguments$width
#> [1] "100%"
#> 
#> 
#> 
#> 
#> 
#> $ui_tree$uiChildren[[3]]
#> $ui_tree$uiChildren[[3]]$uiName
#> [1] "gridlayout::grid_card_plot"
#> 
#> $ui_tree$uiChildren[[3]]$uiArguments
#> $ui_tree$uiChildren[[3]]$uiArguments$area
#> [1] "distPlot"
#> 
#> 
#> 
#> $ui_tree$uiChildren[[4]]
#> $ui_tree$uiChildren[[4]]$uiName
#> [1] "gridlayout::grid_card_plot"
#> 
#> $ui_tree$uiChildren[[4]]$uiArguments
#> $ui_tree$uiChildren[[4]]$uiArguments$area
#> [1] "bluePlot"
#> 
#> 
#> 
#> 
#> 
#> $loaded_libraries
#> character(0)
#>