Launch shinyuieditor server

Spins up an instance of the shinyuieditor server for building new and editing existing Shiny UIs. The console will be blocked while running the editor.

launch_editor(
  app_loc,
  host = "127.0.0.1",
  port = httpuv::randomPort(),
  shiny_background_port = httpuv::randomPort(),
  remove_namespace = TRUE,
  app_preview = TRUE,
  show_logs = TRUE,
  show_preview_app_logs = TRUE,
  launch_browser = interactive(),
  stop_on_browser_close = TRUE
)

Arguments

app_loc

Path to directory containing Shiny app to be visually edited (either containing an app.R or both a ui.R and server.R). If the provided location doesn't exist, or doesn't contain an app, the user will be prompted to choose a starter template which will then be written to the location specified.

host

A string that is a valid IPv4 address that is owned by this server, or "0.0.0.0" to listen on all IP addresses.

port

A number or integer that indicates the server port that should be listened on. Note that on most Unix-like systems including Linux and Mac OS X, port numbers smaller than 1025 require root privileges.

shiny_background_port

Port to launch the shiny app preview on. Typically not necessary to set manually.

remove_namespace

Should namespaces (library:: prefixes) be stripped from the generated UI code? Set to FALSE if you prefer the style of shiny::sliderInput() to sliderInput(). If set to TRUE, then any libraries needed for the nodes will be loaded at the top of your app.R or ui.R.

app_preview

Should a live version of the Shiny app being edited run and auto-show updates made? You may want to disable this if the app has long-running or processor intensive initialization steps.

show_logs

Print status messages to the console?

show_preview_app_logs

Should the logged output of the app preview be printed? Useful for debugging an app that's not working properly. These logs are already shown in the app-preview pane of the editor.

launch_browser

Should the browser be automatically opened to the editor?

stop_on_browser_close

Should the editor server end when the browser window is closed or is dormant for too long. Set this to false if you want to try running the app in a different browser or refreshing the browser etc..

Examples

if (FALSE) {
  # Start editor on a non-existing app directory to choose a starter template
  launch_editor(app_loc = "empty_directory/")

  # You can control where the app runs just like a normal Shiny app.
  # This can be useful if you want to access it from a headless server
  launch_editor(
    app_loc = "my-app/",
    host = "0.0.0.0",
    port = 8888,
    launch_browser = FALSE,
    stop_on_browser_close = FALSE
  )
}